Migrate sentiment classification from nltk to TextBlob
