--- data_processing.py
@@ -1,58 +1,0 @@
-import random, re, string
-
-from nltk import FreqDist, classify, NaiveBayesClassifier
-from nltk.corpus import twitter_samples, stopwords
-from nltk.stem.wordnet import WordNetLemmatizer
-from nltk.tag import pos_tag
-from nltk.tokenize import word_tokenize
-
-
-def remove_noise_and_lemmatize(tweet_tokens, stop_words = ()):
-    cleaned_tokens = []
-    for token, tag in pos_tag(tweet_tokens):
-        # Remove hyperlinks
-        token = re.sub('http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+#]|[!*\(\),]|'\
-                       '(?:%[0-9a-fA-F][0-9a-fA-F]))+','', token)
-        
-        # Remove any twitter @ mentions
-        token = re.sub("(@[A-Za-z0-9_]+)","", token)
-
-        if tag.startswith("NN"):
-            pos = 'n'
-        elif tag.startswith('VB'):
-            pos = 'v'
-        else:
-            pos = 'a'
-
-        lemmatizer = WordNetLemmatizer()
-        token = lemmatizer.lemmatize(token, pos)
-
-        if len(token) > 0 and token not in string.punctuation and token.lower() not in stop_words:
-            cleaned_tokens.append(token.lower())
-
-    return cleaned_tokens
-
-
-def get_all_words(cleaned_tokens_list):
-    for tokens in cleaned_tokens_list:
-        for token in tokens:
-            yield token
-
-
-def get_token_dict_for_model(cleaned_tokens_list):
-    for tweet_tokens in cleaned_tokens_list:
-        yield dict([token, True] for token in tweet_tokens)
-
-
-def create_dataset(raw_tokens, data_label):
-    stop_words = stopwords.words('english')
-
-    cleaned_tokens = []
-    for tokens in raw_tokens:
-        cleaned_tokens.append(remove_noise_and_lemmatize(tokens, stop_words))
-
-    token_dicts = get_token_dict_for_model(cleaned_tokens)
-
-    dataset = [(tweet_dict, data_label) for tweet_dict in token_dicts]
-    return dataset
-
